optimized `Message` routing
